optimized data types

+++ src/create.h
@@ -3,10 +3,10 @@
 
 void create_adj(struct Adjacency * adj, struct Graph * graph) {
   adj->dim = graph->num_nodes;
-  adj->matrix = malloc(adj->dim * sizeof(int*));
+  adj->matrix = malloc(adj->dim * sizeof(unsigned int*));
   for (size_t i = 0; i < adj->dim; i++) {
-    adj->matrix[i] = malloc(adj->dim * sizeof(int));
-    memset(adj->matrix[i], 0, adj->dim * sizeof(int));
+    adj->matrix[i] = malloc(adj->dim * sizeof(unsigned int));
+    memset(adj->matrix[i], 0, adj->dim * sizeof(unsigned int));
   }
   for (size_t i = 0; i < graph->num_nodes; i++)
     for (size_t j = 0; j < graph->num_outgoing[i]; j++)
@@ -24,7 +24,7 @@ int create_graph(char * filename, struct Graph * graph) {
   graph->num_nodes = 0;
   for (size_t i = 0; i < 5; i++)
     getline(&line, &len, handle);
-  sscanf(line, "%u", &graph->num_nodes);
+  sscanf(line, "%hu", &graph->num_nodes);
   graph->id2name = malloc(graph->num_nodes * sizeof(char *));
   for (size_t i = 0; i < graph->num_nodes; i++) {
     /* parse node */
@@ -36,22 +36,22 @@ int create_graph(char * filename, struct Graph * graph) {
   }
   /* get number of edges */
   getline(&line, &len, handle);
-  sscanf(line, "%d", &graph->num_edges);
-  graph->outgoing_edges = malloc(graph->num_nodes * sizeof(int *));
-  graph->incoming_edges = malloc(graph->num_nodes * sizeof(int *));
-  graph->num_outgoing = malloc(graph->num_nodes * sizeof(int *));
-  graph->num_incoming = malloc(graph->num_nodes * sizeof(int *));
-  memset(graph->num_outgoing, 0, graph->num_nodes * sizeof(int *));
-  memset(graph->num_incoming, 0, graph->num_nodes * sizeof(int *));
+  sscanf(line, "%u", &graph->num_edges);
+  graph->outgoing_edges = malloc(graph->num_nodes * sizeof(short int *));
+  graph->incoming_edges = malloc(graph->num_nodes * sizeof(short int *));
+  graph->num_outgoing = malloc(graph->num_nodes * sizeof(unsigned int *));
+  graph->num_incoming = malloc(graph->num_nodes * sizeof(unsigned int *));
+  memset(graph->num_outgoing, 0, graph->num_nodes * sizeof(unsigned int *));
+  memset(graph->num_incoming, 0, graph->num_nodes * sizeof(unsigned int *));
   for (size_t i = 0; i < graph->num_nodes; i++)
     graph->outgoing_edges[i] = malloc(0);
   for (size_t i = 0; i < graph->num_edges; i++) {
     /* parse edge */
     getline(&line, &len, handle);
-    int tail, head;
-    sscanf(line, "%d %d", &tail, &head);
-    size_t tail_size = ++graph->num_outgoing[--tail] * sizeof(int);
-    size_t head_size = ++graph->num_incoming[--head] * sizeof(int);
+    short int tail, head;
+    sscanf(line, "%hu %hu", &tail, &head);
+    size_t tail_size = ++graph->num_outgoing[--tail] * sizeof(short int);
+    size_t head_size = ++graph->num_incoming[--head] * sizeof(short int);
     graph->outgoing_edges[tail] = realloc(graph->outgoing_edges[tail], tail_size);
     graph->incoming_edges[head] = realloc(graph->incoming_edges[head], head_size);
     graph->outgoing_edges[tail][graph->num_outgoing[tail] - 1] = head;
@@ -76,10 +76,10 @@ void create_alignment(char * files[]) {
   }
   create_adj(A1, G1);
   create_adj(A2, G2);
-  G1->translate = malloc(G1->num_nodes * sizeof(int));
-  G2->taken = malloc(G2->num_nodes * sizeof(int));
-  memset(G2->taken, 0, G2->num_nodes * sizeof(int));
-  for (int i = 0; i < G1->num_nodes; i++) {
+  G1->translate = malloc(G1->num_nodes * sizeof(short int));
+  G2->taken = malloc(G2->num_nodes * sizeof(short int));
+  memset(G2->taken, 0, G2->num_nodes * sizeof(short int));
+  for (size_t i = 0; i < G1->num_nodes; i++) {
     G1->translate[i] = i;
     G2->taken[i] = 1;
   }
